EDITED server, database

+++ database/models/payment.js
@@ -1,8 +1,8 @@
 module.exports = (sql,  Sequelize) =>{
     const payment = sql.define('payment',{
         type: Sequelize.STRING,
-        amount: Sequelize.INTEGER,
-        card_no: Sequelize.INTEGER,
+        amount: Sequelize.FLOAT,
+        card_no: Sequelize.INT,
         payment_time: Sequelize.DATE,
 
     });
